codebuild_learn/pyproject.toml
[project]
name = "codebuild-learn"
version = "0.1.0"
description = ""
authors = [
    { name = "ShiYue", email = "yue.shi@amptalk.co.jp" }
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "aws-cdk-lib",
    "constructs>=10.0.0,<11.0.0",
    "boto3 (>=1.37.1,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "codebuild_learn", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"